+++ include/bbque/app/working_mode_status.h
@@ -96,7 +96,7 @@ public:
 	 * @brief Working Mode ID
 	 * @return An integer number
 	 */
-	virtual int8_t Id() const = 0;
+	virtual int32_t Id() const = 0;
 
 	/**
 	 * @brief Get the application owning the working mode
